+++ backend/src/main/java/cn/zl/account/book/controller/response/FundsRecordTopResponse.java
@@ -21,7 +21,7 @@ public class FundsRecordTopResponse {
     @Builder
     public static class TopInfo {
 
-        private Long fundsRecordBalance;
+        private Double fundsRecordBalance;
 
         private LocalDateTime fundsRecordTime;
